Safari-611.1.10/LayoutTests/js/instance-property-getter-other-instance-expected.txt

Tests that calling an instance property getter on another instance works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS locationGetter is an instance of function Function() {
    [native code]
}
PASS locationGetter.call(otherWindow.document).toString() is "about:blank"
PASS locationGetter.call(window.document) === window.document.location is true
PASS successfullyParsed is true

TEST COMPLETE
